REFACTOR: moved user entity and repo into respective folders
