+++ backend/src/main/java/com/skysocial/backend/repos/UserRepo.java
@@ -1,5 +1,6 @@
-package com.skysocial.backend.entities;
+package com.skysocial.backend.repos;
 
+import com.skysocial.backend.entities.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
